+++ RuntimeEditor/Utils/GUIDUtils.cs
@@ -8,11 +8,15 @@ namespace Hananoki {
 
 	public static class GUIDUtils {
 		public static T LoadAssetAtGUID<T>( string guid ) where T : UnityEngine.Object {
-			var path = UnityEditor.AssetDatabase.GUIDToAssetPath( guid );
+			var path = AssetDatabase.GUIDToAssetPath( guid );
 			T b = AssetDatabase.LoadAssetAtPath( path, typeof( T ) ) as T;
 			return b;
 		}
-
+		public static UnityEngine.Object LoadAssetAtGUID( string guid ) {
+			var path = AssetDatabase.GUIDToAssetPath( guid );
+			UnityEngine.Object b = AssetDatabase.LoadAssetAtPath( path, typeof( UnityEngine.Object ) ) as UnityEngine.Object;
+			return b;
+		}
 
 		public static string GetAssetPath( string guid ) {
 			return AssetDatabase.GUIDToAssetPath( guid );
@@ -33,7 +37,9 @@ namespace Hananoki {
 
 
 		public static string AdjustDirectoryGUID( this string guid ) {
-			var path = GUIDUtils.GetAssetPath( guid );
+			if( guid.IsEmpty() ) return string.Empty;
+
+			var path = GetAssetPath( guid );
 			if( Directory.Exists( path ) ) {
 				return guid;
 			}
@@ -41,6 +47,15 @@ namespace Hananoki {
 				return path.DirectoryName().ToGUID();
 			}
 		}
+
+		public static string NewString() {
+			return System.Guid.NewGuid().ToString( "N" );
+		}
+
+		public static int NewInt32() {
+			return System.Convert.ToInt32( NewString().Substring( 0, 8 ), 16 );
+		}
+
 	}
 }